fix(exp6/b0): soft-label CE bf16 dtype mismatch (autocast scatter_ failure)
First HF Jobs training run failed with:
  RuntimeError: scatter(): Expected self.dtype to be equal to src.dtype
  losses.py:217 _soft_label_ce_masked

Under bf16 autocast, head_logits are bfloat16 → soft_target = zeros_like(logits)
is bf16. But lower_w / upper_w were computed from fp32 expert_continuous and
fp32 bin_centers, so they stayed fp32. scatter_ requires matching dtypes.

Fix: cast lower_w / upper_w to logits.dtype before scatter. Applied to both
soft_label_ce (unmasked) and _soft_label_ce_masked (used by total_loss_b0).

Added 2 regression tests with bf16 logits to prevent silent regression — local
tests use fp32 by default so this would have slipped past the existing suite.

+++ experiments/action_primitives/losses.py
@@ -131,8 +131,10 @@ def soft_label_ce(
     span = torch.where(span > 1e-6, span, torch.ones_like(span))
     upper_w = (expert_continuous - lower_centers) / span
     lower_w = 1.0 - upper_w
-    upper_w = torch.clamp(upper_w, 0.0, 1.0)
-    lower_w = torch.clamp(lower_w, 0.0, 1.0)
+    # Cast weights to logits.dtype so scatter_ works under bf16 autocast
+    # (logits are bf16; weights computed from fp32 expert_continuous + bin_centers).
+    upper_w = torch.clamp(upper_w, 0.0, 1.0).to(logits.dtype)
+    lower_w = torch.clamp(lower_w, 0.0, 1.0).to(logits.dtype)
 
     # Construct soft target (B, num_bins)
     soft_target = torch.zeros_like(logits)
@@ -211,8 +213,10 @@ def _soft_label_ce_masked(
     upper_centers = bin_centers[upper_idx]
     span = upper_centers - lower_centers
     span = torch.where(span > 1e-6, span, torch.ones_like(span))
-    upper_w = torch.clamp((expert_continuous - lower_centers) / span, 0.0, 1.0)
-    lower_w = 1.0 - upper_w
+    # Cast weights to logits.dtype so scatter_ works under bf16 autocast
+    # (logits are bf16; weights computed from fp32 expert_continuous + bin_centers).
+    upper_w = torch.clamp((expert_continuous - lower_centers) / span, 0.0, 1.0).to(logits.dtype)
+    lower_w = (1.0 - upper_w).to(logits.dtype)
     soft_target = torch.zeros_like(logits)
     soft_target.scatter_(-1, lower_idx.unsqueeze(-1), lower_w.unsqueeze(-1))
     soft_target.scatter_add_(-1, upper_idx.unsqueeze(-1), upper_w.unsqueeze(-1))
